whisper::agent::session

Trait ResponseSink

pub trait ResponseSink {
    // Required methods
    fn chat_start(
        &mut self,
    ) -> Pin<Box<dyn Future<Output = Result<(), SinkError>> + Send + '_>>;
    fn user_start(
        &mut self,
    ) -> Pin<Box<dyn Future<Output = Result<(), SinkError>> + Send + '_>>;
    fn output_start(
        &mut self,
    ) -> Pin<Box<dyn Future<Output = Result<(), SinkError>> + Send + '_>>;
    fn output_text(
        &mut self,
        content: &(dyn Display + Send + Sync),
    ) -> Pin<Box<dyn Future<Output = Result<(), SinkError>> + Send + '_>>;
    fn output_reason_start(
        &mut self,
    ) -> Pin<Box<dyn Future<Output = Result<(), SinkError>> + Send + '_>>;
    fn output_reason_end(
        &mut self,
    ) -> Pin<Box<dyn Future<Output = Result<(), SinkError>> + Send + '_>>;
    fn output_finished(
        &mut self,
        usage: &Option<Usage>,
    ) -> Pin<Box<dyn Future<Output = Result<(), SinkError>> + Send + '_>>;
    fn chat_finished(
        &mut self,
    ) -> Pin<Box<dyn Future<Output = Result<(), SinkError>> + Send + '_>>;
    fn output_error(
        &mut self,
        e: &(dyn Display + Send + Sync),
    ) -> Pin<Box<dyn Future<Output = Result<(), SinkError>> + Send + '_>>;
}

Trait to abstract display

Required Methods

fn chat_start( &mut self, ) -> Pin<Box<dyn Future<Output = Result<(), SinkError>> + Send + '_>>

Output the string to indicate the start of the chat

fn user_start( &mut self, ) -> Pin<Box<dyn Future<Output = Result<(), SinkError>> + Send + '_>>

Output the string to indicate the start of user’s query

fn output_start( &mut self, ) -> Pin<Box<dyn Future<Output = Result<(), SinkError>> + Send + '_>>

Output the string to indicate the start of assistant’s answer

fn output_text( &mut self, content: &(dyn Display + Send + Sync), ) -> Pin<Box<dyn Future<Output = Result<(), SinkError>> + Send + '_>>

Output the normal text

fn output_reason_start( &mut self, ) -> Pin<Box<dyn Future<Output = Result<(), SinkError>> + Send + '_>>

Output the start of reasoning content

fn output_reason_end( &mut self, ) -> Pin<Box<dyn Future<Output = Result<(), SinkError>> + Send + '_>>

Output the end of reasoning content

fn output_finished( &mut self, usage: &Option<Usage>, ) -> Pin<Box<dyn Future<Output = Result<(), SinkError>> + Send + '_>>

Output the end of assistant’s answer

fn chat_finished( &mut self, ) -> Pin<Box<dyn Future<Output = Result<(), SinkError>> + Send + '_>>

Output the end of chat

fn output_error( &mut self, e: &(dyn Display + Send + Sync), ) -> Pin<Box<dyn Future<Output = Result<(), SinkError>> + Send + '_>>

Output the error